[ETC] 코드 정리

+++ src/main/java/com/wooyeon/yeon/likeManage/repository/LikeRepository.java
@@ -16,7 +16,7 @@ import java.util.UUID;
 public interface LikeRepository extends JpaRepository<UserLike, Long>, LikeRepositoryFindProfileList {
     @Query("select count(userLike) from UserLike userLike " +
             "where (userLike.likeFrom.userId = :fromId and userLike.likeTo.userId = :toId) ")
-    long countMatch(@Param("fromId") Long fromId, @Param("toId") Long toId);
+    int countMatch(@Param("fromId") Long fromId, @Param("toId") Long toId);
 
     /**
      * 해당하는 user_like를 찾아줌.
